Fix letterRoutes and log letter attempt
letterRoutes: Commented out the unwanted route. Multiple routes with the same method and path is not allowed in Express.
Log Letter Attempt: The previous active log letter attempt function expected audio, absence of which causes the code to crash. Hence, modified it in such a way that audio is optional.